Rename GeometricComputationsTests to match other test projects.
