+++ DotNet/GeometricComputationsTests/GeometricComputationsTests.csproj
@@ -8,8 +8,8 @@
     <ProjectGuid>{A9F2ADF7-4447-40B2-8975-9FCB72C6694F}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>TestGeometricComputations</RootNamespace>
-    <AssemblyName>TestGeometricComputations</AssemblyName>
+    <RootNamespace>GeometricComputationsTests</RootNamespace>
+    <AssemblyName>GeometricComputationsTests</AssemblyName>
     <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
     <TargetFrameworkProfile>
     </TargetFrameworkProfile>
