+++ DotNet/GeometricComputationsTests/Properties/AssemblyInfo.cs
